requirements.txt
altgraph==0.17.4
packaging==25.0
pefile==2023.2.7
pillow==12.0.0
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pyserial==3.5
PySide6==6.10.0
PySide6_Addons==6.10.0
PySide6_Essentials==6.10.0
pywin32-ctypes==0.2.3
setuptools==80.9.0
shiboken6==6.10.0
